+++ src/main/java/net/chaos/chaosmod/blocks/BlockContainerBase.java
@@ -4,18 +4,15 @@ import net.chaos.chaosmod.init.ModBlocks;
 import net.chaos.chaosmod.init.ModItems;
 import net.minecraft.block.BlockContainer;
 import net.minecraft.block.ITileEntityProvider;
-import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
 import proxy.IBlockModel;
 
 public abstract class BlockContainerBase extends BlockContainer implements IBlockModel, ITileEntityProvider {
+
 	protected BlockContainerBase(String name, Material materialIn) {
 		super(materialIn);
 		setUnlocalizedName(name);
 		setRegistryName(name);
-		setHardness(2.0F);
-		setResistance(5.0F);
-		setSoundType(SoundType.WOOD);
 	
 		ModBlocks.BLOCKS.add(this);
 		ModItems.ITEMS.add(new ItemBlockBase(this).setRegistryName(this.getRegistryName()));