+++ src/main/java/net/chaos/chaosmod/init/ModBiomes.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import net.chaos.chaosmod.biomes.BiomeDenseForest;
+import net.chaos.chaosmod.biomes.BiomeSpring;
 import net.chaos.chaosmod.world.gen.experimental.biomes.ExperimentalBiome;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.biome.Biome;
@@ -19,21 +20,22 @@ public class ModBiomes {
 	public static Biome DENSE_FOREST = new BiomeDenseForest().setRegistryName(new ResourceLocation(Reference.MODID, "dense_forest"));
 //	public static Biome NETHER_CAVES = new CustomBiomeNether().setRegistryName(new ResourceLocation(Reference.MODID, "nether_caves"));
 //	public static Biome ENDER_GARDEN = new CustomBiomeEnderGarden().setRegistryName(new ResourceLocation(Reference.MODID, "ender_garden"));
-//	public static Biome CHAOS_LAND_BIOME = new CustomBiomeChaosLand().setRegistryName(new ResourceLocation(Reference.MODID, "chaos_land_biome"));
+	public static Biome SPRING_BIOME = new BiomeSpring().setRegistryName(new ResourceLocation(Reference.MODID, "spring_biome"));
 //	public static Biome CUSTOM_HELL = new CustomBiomeHell().setRegistryName(new ResourceLocation(Reference.MODID, "custom_hell"));
 	public static Biome EXPERIMENTAL = new ExperimentalBiome().setRegistryName(new ResourceLocation(Reference.MODID, "experimental_biome"));
 
     public static void registerBiomes() {
     	BIOMES.add(DENSE_FOREST);
+        BIOMES.add(SPRING_BIOME);
 //        BIOMES.add(NETHER_CAVES);
 //        BIOMES.add(ENDER_GARDEN);
-//        BIOMES.add(CHAOS_LAND_BIOME);
 //        BIOMES.add(CUSTOM_HELL);
         BIOMES.add(EXPERIMENTAL);
     }
     
     public static void init() {
     	addBiome(ModBiomes.DENSE_FOREST, 10, BiomeType.WARM, BiomeDictionary.Type.FOREST, BiomeDictionary.Type.CONIFEROUS, BiomeDictionary.Type.DENSE);
+    	addBiome(ModBiomes.SPRING_BIOME, 10, BiomeType.WARM, BiomeDictionary.Type.PLAINS, BiomeDictionary.Type.FOREST, BiomeDictionary.Type.HOT, BiomeDictionary.Type.PLAINS);
 //        addBiome(ModBiomes.NETHER_CAVES);
 //        addBiome(ModBiomes.ENDER_GARDEN);
 //        addBiome(ModBiomes.CHAOS_LAND_BIOME);