+++ src/main/java/util/Reference.java
@@ -29,6 +29,7 @@ public class Reference {
 	public static final int GUI_MARKET_ID = 9;
 	public static final int GUI_FISHINGGAME_ID = 10;
 	public static final int GUI_BEAR_ID = 11;
+	public static final int GUI_JIGSAW_ID = 12;
 
     // Entities ids
     private static int entityId = 120;